RemoveBaseUri testing and resulting bugfixes
git-svn-id: https://uriparser.svn.sourceforge.net/svnroot/uriparser/uriparser/trunk@239 119732b9-2324-0410-a8a5-815c10060bbe

+++ lib/UriFile.c
@@ -85,7 +85,7 @@ static URI_INLINE int URI_FUNC(FilenameToUriString)(const URI_CHAR * filename,
 			if (lastSep + 1 < input) {
 				if (!unix && (firstSegment == URI_TRUE)) {
 					/* Quick hack to not convert "C:" to "C%3A" */
-					const int charsToCopy = input - (lastSep + 1);
+					const int charsToCopy = (int)(input - (lastSep + 1));
 					memcpy(output, lastSep + 1, charsToCopy * sizeof(URI_CHAR));
 					output += charsToCopy;
 				} else {
